fix MaxOffersPerEstateType

+++ pallets/estate/src/mock.rs
@@ -459,7 +459,7 @@ parameter_types! {
 	pub const RewardPaymentDelay: u32 = 2;
 	pub const DefaultMaxBound: (i32,i32) = MAX_BOUND;
 	pub const NetworkFee: Balance = 1; // Network fee
-	pub const MaxOffersPerEstate: u32 = 2;
+	pub const MaxOffersPerEstate: usize = 2;
 	pub const MinLeasePricePerBlock: Balance = 1u128;
 	pub const MaxLeasePeriod: u32 = 9;
 	pub const LeaseOfferExpiryPeriod: u32 = 6;
